pyproject.toml
[project]
name = "ghostroot"
version = "0.1.0"
description = "Agentic conlang archaeology PoC"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
  "rich>=13.0.0"
]

[project.optional-dependencies]
openai = [
  "openai>=1.0.0",
  "openai-agents>=0.0.1"
]

[project.scripts]
ghostroot = "ghostroot.run:main"

[tool.uv]
dev-dependencies = [
  "pytest>=8.0.0",
  "ruff>=0.6.0"
]

[tool.ruff]
line-length = 100
